Moved Object Image Exportation to separate class. Added ability to use morphological volume if it exists.
